move DateFilter for Sphinx to separate class
